BigDecimal fix

+++ src/main/java/com/example/project_1/repository/EventRepository.java
@@ -7,6 +7,8 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -14,5 +16,5 @@ public interface EventRepository extends JpaRepository<Event, Long> {
     List<Event> findByStatus(EventStatus status);
 
     @Query("SELECT SUM(tt.price) FROM Booking b JOIN b.ticketType tt WHERE tt.event.event_id = :eventId AND b.paymentStatus = 'CONFIRMED'")
-    Double calculateConfirmedRevenue(@Param("eventId") Long eventId);
+    BigDecimal calculateConfirmedRevenue(@Param("eventId") Long eventId);
 }
